ref: remove OptionMixin
this fixes types for the various models involved
