+++ src/sentry/models/organization.py
@@ -28,11 +28,11 @@ from sentry.db.models import (
     sane_repr,
 )
 from sentry.db.models.fields.slug import SentryOrgSlugField
+from sentry.db.models.manager import ValidateFunction
 from sentry.db.models.outboxes import ReplicatedRegionModel
 from sentry.db.models.utils import slugify_instance
 from sentry.db.postgres.transactions import in_test_hide_transaction_boundary
 from sentry.locks import locks
-from sentry.models.options.option import OptionMixin
 from sentry.models.outbox import OutboxCategory
 from sentry.roles.manager import Role
 from sentry.services.hybrid_cloud.notifications import notifications_service
@@ -144,7 +144,7 @@ class OrganizationManager(BaseManager["Organization"]):
 
 @snowflake_id_model
 @region_silo_model
-class Organization(ReplicatedRegionModel, OptionMixin, OrganizationAbsoluteUrlMixin):
+class Organization(ReplicatedRegionModel, OrganizationAbsoluteUrlMixin):
     """
     An organization represents a group of individuals which maintain ownership of projects.
     """
@@ -476,3 +476,14 @@ class Organization(ReplicatedRegionModel, OptionMixin, OrganizationAbsoluteUrlMi
         if not self.get_option("sentry:alerts_member_write", ALERTS_MEMBER_WRITE_DEFAULT):
             scopes.discard("alerts:write")
         return frozenset(scopes)
+
+    def get_option(
+        self, key: str, default: Any | None = None, validate: ValidateFunction | None = None
+    ) -> Any:
+        return self.option_manager.get_value(self, key, default, validate)
+
+    def update_option(self, key: str, value: Any) -> bool:
+        return self.option_manager.set_value(self, key, value)
+
+    def delete_option(self, key: str) -> None:
+        self.option_manager.unset_value(self, key)
